move constructor logic
