+++ netket/optimizer/qgt/qgt_jacobian_dense.py
@@ -13,152 +13,18 @@
 # limitations under the License.
 
 from typing import Optional, Union
-from functools import partial
 
 import jax
 from jax import numpy as jnp
 from flax import struct
 
 from netket.utils.types import Scalar, PyTree
-from netket.utils import mpi, timing
-import netket.jax as nkjax
-from netket.nn import split_array_mpi
+from netket.utils import mpi
+from netket import jax as nkjax
 
 from ..linear_operator import LinearOperator, Uninitialized
 
 from .common import check_valid_vector_type
-from .qgt_jacobian_common import (
-    sanitize_diag_shift,
-    to_shift_offset,
-    rescale,
-)
-
-
-@timing.timed
-def QGTJacobianDense(
-    vstate=None,
-    *,
-    mode: Optional[str] = None,
-    holomorphic: Optional[bool] = None,
-    diag_shift=None,
-    diag_scale=None,
-    rescale_shift=None,
-    chunk_size=None,
-    **kwargs,
-) -> "QGTJacobianDenseT":
-    """
-    Semi-lazy representation of an S Matrix where the Jacobian O_k is precomputed
-    and stored as a dense matrix.
-
-    The matrix of gradients O is computed on initialisation, but not S,
-    which can be computed by calling :code:`to_dense`.
-    The details on how the ⟨S⟩⁻¹⟨F⟩ system is solved are contained in
-    the field `sr`.
-
-    Numerical estimates of the QGT are usually ill-conditioned and require
-    regularisation. The standard approach is to add a positive constant to the diagonal;
-    alternatively, Becca and Sorella (2017) propose scaling this offset with the
-    diagonal entry itself. NetKet allows using both in tandem:
-
-    .. math::
-
-        S_{ii} \\mapsto S_{ii} + \\epsilon_1 S_{ii} + \\epsilon_2;
-
-    :math:`\\epsilon_{1,2}` are specified using `diag_scale` and `diag_shift`,
-    respectively.
-
-    Args:
-        vstate: The variational state
-        mode: "real", "complex" or "holomorphic": specifies the implementation
-              used to compute the jacobian. "real" discards the imaginary part
-              of the output of the model. "complex" splits the real and imaginary
-              part of the parameters and output. It works also for non holomorphic
-              models. holomorphic works for any function assuming it's holomorphic
-              or real valued.
-        holomorphic: a flag to indicate that the function is holomorphic.
-        diag_scale: Fractional shift :math:`\\epsilon_1` added to diagonal entries (see above).
-        diag_shift: Constant shift :math:`\\epsilon_2` added to diagonal entries (see above).
-        chunk_size: If supplied, overrides the chunk size of the variational state
-                    (useful for models where the backward pass requires more
-                    memory than the forward pass).
-    """
-    if mode is not None and holomorphic is not None:
-        raise ValueError("Cannot specify both `mode` and `holomorphic`.")
-    if rescale_shift is not None and diag_scale is not None:
-        raise ValueError("Cannot specify both `rescale_shift` and `diag_scale`.")
-
-    if vstate is None:
-        return partial(
-            QGTJacobianDense,
-            mode=mode,
-            holomorphic=holomorphic,
-            chunk_size=chunk_size,
-            diag_shift=diag_shift,
-            diag_scale=diag_scale,
-            rescale_shift=rescale_shift,
-            **kwargs,
-        )
-
-    diag_shift, diag_scale = sanitize_diag_shift(diag_shift, diag_scale, rescale_shift)
-
-    # TODO: Find a better way to handle this case
-    from netket.vqs import FullSumState
-
-    if isinstance(vstate, FullSumState):
-        samples = split_array_mpi(vstate._all_states)
-        pdf = split_array_mpi(vstate.probability_distribution())
-    else:
-        samples = vstate.samples
-        if samples.ndim >= 3:
-            # use jit so that we can do it on global shared array
-            samples = jax.jit(jax.lax.collapse, static_argnums=(1, 2))(samples, 0, 2)
-        pdf = None
-
-    if mode is None:
-        mode = nkjax.jacobian_default_mode(
-            vstate._apply_fun,
-            vstate.parameters,
-            vstate.model_state,
-            samples,
-            holomorphic=holomorphic,
-        )
-
-    if chunk_size is None and hasattr(vstate, "chunk_size"):
-        chunk_size = vstate.chunk_size
-
-    shift, offset = to_shift_offset(diag_shift, diag_scale)
-
-    jacobians = nkjax.jacobian(
-        vstate._apply_fun,
-        vstate.parameters,
-        samples,
-        vstate.model_state,
-        mode=mode,
-        pdf=pdf,
-        chunk_size=chunk_size,
-        dense=True,
-        center=True,
-        _sqrt_rescale=True,
-    )
-
-    if offset is not None:
-        ndims = 1 if mode != "complex" else 2
-        jacobians, scale = rescale(jacobians, offset, ndims=ndims)
-    else:
-        scale = None
-
-    pars_struct = jax.tree_util.tree_map(
-        lambda x: jax.ShapeDtypeStruct(x.shape, x.dtype), vstate.parameters
-    )
-
-    return QGTJacobianDenseT(
-        O=jacobians,
-        scale=scale,
-        mode=mode,
-        _params_structure=pars_struct,
-        diag_shift=shift,
-        **kwargs,
-    )
 
 
 @struct.dataclass
